+++ TorrServer/src/torrentserver/storage/state/state.go
@@ -6,9 +6,9 @@ import (
 
 type CacheState struct {
 	Hash          string
-	Capacity      int
-	Filled        int
-	PiecesLength  int
+	Capacity      int64
+	Filled        int64
+	PiecesLength  int64
 	PiecesCount   int
 	CurrentRead   int
 	EndRead       int
@@ -19,7 +19,7 @@ type CacheState struct {
 type ItemState struct {
 	Id         int
 	Accessed   time.Time
-	BufferSize int
+	BufferSize int64
 	Completed  bool
 	Hash       string
 }
